Updates to accommodate move of api/config/mcp to api/mcp. (#6835)
* Updates to accommodate move of api/config/mcp to api/mcp.

* Update dependencies.

* Updated dep and reran due to linter problem. Reverted license file
updates.
