+++ Gopkg.lock
@@ -1291,14 +1291,14 @@
   version = "v2.2.1"
 
 [[projects]]
-  branch = "master"
+  branch = "release-1.0"
   name = "istio.io/api"
   packages = [
     "authentication/v1alpha1",
     "broker/dev",
-    "config/mcp/v1alpha1",
     "envoy/config/filter/http/authn/v2alpha1",
     "envoy/config/filter/http/jwt_auth/v2alpha1",
+    "mcp/v1alpha1",
     "mesh/v1alpha1",
     "mixer/adapter/model/v1beta1",
     "mixer/v1",
@@ -1308,7 +1308,7 @@
     "rbac/v1alpha1",
     "routing/v1alpha1"
   ]
-  revision = "2d182f5d150eaa9ac276347df8f0154f209973dc"
+  revision = "80b4a7b8366713fe6295761d7e513403076316a4"
 
 [[projects]]
   name = "istio.io/fortio"
@@ -1606,6 +1606,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "70f66277068be52e07b9b2e40c6959ddae2b8708745af55a9a22cac843b99210"
+  inputs-digest = "ad523c6f8d7a2659cb97f0d6daf0b93da4107d4f340313d94115a6f5b70908fd"
   solver-name = "gps-cdcl"
   solver-version = 1
