+++ galley/pkg/mcp/server/server_test.go
@@ -31,7 +31,7 @@ import (
 	"google.golang.org/grpc/peer"
 	"google.golang.org/grpc/status"
 
-	mcp "istio.io/api/config/mcp/v1alpha1"
+	mcp "istio.io/api/mcp/v1alpha1"
 )
 
 type mockConfigWatcher struct {
