+++ galley/pkg/mcp/snapshot/inmemory_test.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/gogo/protobuf/types"
 
-	mcp "istio.io/api/config/mcp/v1alpha1"
+	mcp "istio.io/api/mcp/v1alpha1"
 )
 
 func TestNewInMemory(t *testing.T) {
