+++ galley/pkg/mcp/snapshot/snapshot.go
@@ -18,7 +18,7 @@ import (
 	"sync"
 	"time"
 
-	mcp "istio.io/api/config/mcp/v1alpha1"
+	mcp "istio.io/api/mcp/v1alpha1"
 	"istio.io/istio/galley/pkg/mcp/server"
 	"istio.io/istio/pkg/log"
 )
