+++ galley/pkg/server/server.go
@@ -22,7 +22,7 @@ import (
 
 	"google.golang.org/grpc"
 
-	mcp "istio.io/api/config/mcp/v1alpha1"
+	mcp "istio.io/api/mcp/v1alpha1"
 	"istio.io/istio/galley/pkg/kube"
 	"istio.io/istio/galley/pkg/mcp/server"
 	"istio.io/istio/galley/pkg/mcp/snapshot"
