+++ galley/tools/mcpc/main.go
@@ -23,7 +23,7 @@ import (
 
 	"google.golang.org/grpc"
 
-	mcp "istio.io/api/config/mcp/v1alpha1"
+	mcp "istio.io/api/mcp/v1alpha1"
 	"istio.io/istio/galley/pkg/mcp/client"
 
 	// Import the resource package to pull in all proto types.
